video-editor-pro/scripts/requirements.txt
torch
torchaudio
openai-whisper
ffmpeg-python
numpy
mediapipe
opencv-python
assemblyai\npython-dotenv
